remove GenericMetadataRepository interface and rename implementation
